\name{stepp.rnote}
\alias{stepp.rnote}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  The method to print the release note for STEPP.
}

\description{
  This method will print the current stepp version and the corresponding release note.
}

\usage{
  stepp.rnote()
}


\author{
 Wai-ki Yip, Marco Bonetti, Bernard Cole, Ann Lazar, Victoria Xin Wang, Richard Gelber
}


\examples{
  stepp.rnote()
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
