\name{test}
\alias{test}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  the standard generic function for all test methods
}

\description{
   The default generic function for test methods for all stepp models classes and the steppes class.

   For detail, please refer to the documentation in the test method in the S4 class: stmodelCI, stmodelKM and stmodelGLM.
}


\author{
 Wai-ki Yip
}


\seealso{
  \code{\linkS4class{stwin}},     \code{\linkS4class{stsubpop}},  \code{\linkS4class{stmodelKM}}, 
  \code{\linkS4class{stmodelCI}}, \code{\linkS4class{stmodelGLM}}, 
  \code{\linkS4class{steppes}},   \code{\linkS4class{stmodel}}, 
  \code{\link{stepp.win}},        \code{\link{stepp.subpop}},     \code{\link{stepp.KM}},
  \code{\link{stepp.CI}},         \code{\link{stepp.GLM}},
  \code{\link{generate}},         \code{\link{estimate}}

}
