% --- Source file: stepp_summary.Rd ---
\name{stepp_summary}
\alias{stepp_summary}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Subpopulation Treatment Effect Pattern Plot (STEPP) Accessory Function - stepp_summary
}
\description{
A method to print the summary of the result of a stepp analysis that is captured in the stepp object.   
}
\usage{
stepp_summary(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{the stepp object created using stepp}
}
\details{
  A stepp summary is produced by this method which contains the version information, creation time, execution time, and the subpopulation summary of the stepp analysis.

  See other related stepp functions:
  1. stepp
  2. stepp_summary
  3. stepp_print
  4. stepp_plot
  5. analyze.KM.stepp
  6. analyze.CumInc.stepp
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.


